fixed moving tests to own files under tests/ folder
